requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
dnspython==2.6.1
ecom_pack==0.0.0
email_validator==2.1.1
exceptiongroup==1.2.0
Faker==23.3.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pip==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
wheel==0.42.0